Move dataset loading to client.py; update readme
